refactor: replace 'repository.Collections' with 'mongo.Collections' across repositories and routes
