+++ Delivery/routers/blog_post_routes.go
@@ -15,15 +15,15 @@ func NewBlogRoutes(env *bootstrap.Env, api *gin.RouterGroup, db mongo.Database) 
 	blogGroup := api.Group("/blogs")
 
 	blog_post_controller := controllers.BlogPostController{
-		BlogPostUsecase: usecases.NewBlogPostUsecase(repository.NewBlogPostRepo(db, &repository.Collections{
-			BlogPosts: env.BlogPostCollection,
+		BlogPostUsecase: usecases.NewBlogPostUsecase(repository.NewBlogPostRepo(db, &mongo.Collections{
+			BlogPosts:         env.BlogPostCollection,
 			BlogComments:      env.BlogCommentCollection,
 			BlogUserReactions: env.BlogUserReactionCollection,
 		}), time.Duration(env.CtxTSeconds)*time.Second),
-		Env:             env,
+		Env: env,
 	}
 
-	blogGroup.GET("/", blog_post_controller.GetBlogPosts) // Get all blogs with optional filters
+	blogGroup.GET("/", blog_post_controller.GetBlogPosts)       // Get all blogs with optional filters
 	blogGroup.GET("/:id", blog_post_controller.GetBlogPostByID) // Get a single blog by ID
 	// blogGroup.POST("/", blog_post_controller.CreateBlog) // Create a new blog
 	// blogGroup.PUT("/:id", blog_post_controller.UpdateBlog) // Update an existing blog
