+++ Repositories/blog/blog_ai_repo.go
@@ -8,7 +8,7 @@ import (
 
 type blogAIRepository struct {
 	db          mongo.Database
-	collections *Collections
+	collections *mongo.Collections
 }
 
 // GetGeneratedContentByID implements domain.BlogAIRepository.
@@ -31,7 +31,7 @@ func (b *blogAIRepository) StoreGeneratedContent(ctx context.Context, content *d
 	panic("unimplemented")
 }
 
-func NewBlogAIRepository(db mongo.Database, collections *Collections) domain.BlogAIRepository {
+func NewBlogAIRepository(db mongo.Database, collections *mongo.Collections) domain.BlogAIRepository {
 	return &blogAIRepository{
 		db:          db,
 		collections: collections,
